Enable fx86cl64 to be built on FreeBSD 12.0-RC3
One change is simple: we just need to pick up a definition for
the name T_PAGEFLT.

The other is needed because FreeBSD now uses a new linker by
default (namely the LLVM project's LLD linker).  This linker
gets indigestion, reporting the error

/usr/bin/ld: error: x86-subprims64.o: SHF_MERGE section size must be a multiple of sh_entsize

Fortunately, there's a way to tell the cc compiler driver to
use the older GNU ld, which still works.

See #173.