In acode-var-type, check for $vbitclosed also
If the var was setq'ed, then it is clear that we can't assume that
the var's type is the same type as the initform.

But if the var is closed over, it's also possible for code in some
peer scope to setq the var, so we can't use the type of the initform
in that case either.

At least I think that's what it is.

See #171.